requirements.txt
pdfminer.six==20221105
qdrant-client
sentence-transformers
transformers
torch